+++ NEWS.md
@@ -2,6 +2,7 @@ lmerTest 3.1-3 (October 2020)
 ------------------
 
 - Update a test to accomodate `lme4::anova` reporting `NA` instead of `1` for p-value when df=0.
+- No longer print message about missing cells with mean centered covariates for type III anova tables (issue #19)
 
 lmerTest 3.1-2 (April 2020)
 ------------------